Refactoring filter code
