+++ docs/filtering_and_sorting.md
@@ -7,10 +7,10 @@ To enable filtering on your rest endpoints edit your settings file and modify `R
 REST_FRAMEWORK = {
     ...
     'DEFAULT_FILTER_BACKENDS': (
-        'ansible_base.utils.rest_filters.TypeFilterBackend',
-        'ansible_base.utils.rest_filters.FieldLookupBackend',
+        'ansible_base.filters.rest_framework.type_filter_backend.TypeFilterBackend',
+        'ansible_base.filters.rest_framework.field_lookup_backend.FieldLookupBackend',
         'rest_framework.filters.SearchFilter',
-        'ansible_base.utils.rest_filters.OrderByBackend',
+        'ansible_base.filters.rest_framework.order_backend.OrderByBackend',
     ),
     ...
 }
